fix case where repeat='y'
